third_party/WebKit/LayoutTests/http/tests/inspector/cache-storage/cache-live-update-cache-content-expected.txt

Tests that cache content is marked dirty if it is modified.

Dumping CacheStorage tree:
    (empty)
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        (cache empty)
Cache marked dirty = false
Added entry
Cache marked dirty = true
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        0, http://fake.request.com/1, OK
Cache marked dirty = false
Deleted entry
Cache marked dirty = true
Dumping CacheStorage tree:
    cache: testCache1 - http://127.0.0.1:8000
        (cache empty)
Cache marked dirty = false
